server/yolov9/requirements.txt
absl-py==2.1.0
albumentations==1.4.4
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
backcall==0.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.13.4
flask==3.0.3
Flask-Cors==4.0.0
fonttools==4.51.0
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
google-auth-oauthlib==1.0.0
grpcio==1.62.2
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.34.0
importlib-metadata==7.1.0
importlib-resources==6.4.0
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
kiwisolver==1.4.5
lazy-loader==0.4
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
openai==1.23.2
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==24.0
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
prompt-toolkit==3.0.43
protobuf==5.26.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycocotools==2.0.7
pydantic==2.7.0
pydantic-core==2.18.1
pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.0
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.3
supervision==0.19.0
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
threadpoolctl==3.4.0
tifffile==2023.7.10
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
traitlets==5.14.2
triton==2.2.0
typing-extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
werkzeug==3.0.2
zipp==3.18.1
